Fix DEBT-355 cross-feature question-flow coupling
