+++ app/(app)/app/practice/[sessionId]/hooks/use-practice-session-mark-for-review.ts
@@ -7,11 +7,11 @@ import {
   useRef,
   useState,
 } from 'react';
+import type { LoadState } from '@/app/(app)/app/practice/practice-page-logic';
 import {
   getActionResultErrorMessage,
   getThrownErrorMessage,
-} from '@/app/(app)/app/practice/practice-logic';
-import type { LoadState } from '@/app/(app)/app/practice/practice-page-logic';
+} from '@/app/(app)/app/shared/error-message-helpers';
 import { STANDARD_MUTATION_TIMEOUT_MS } from '@/app/(app)/app/shared/timeout-tiers';
 import {
   reportClientError,
